✨ Feat: #11 학습 완료 API 구현

+++ src/main/java/com/finale/finale/book/repository/UnknownWordRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface UnknownWordRepository extends JpaRepository<UnknownWord, Integer> {
+public interface UnknownWordRepository extends JpaRepository<UnknownWord, Long> {
 
     @Query("SELECT uw FROM UnknownWord uw WHERE uw.user.id = :userId AND uw.nextReviewDate <= :today ORDER BY uw.nextReviewDate ASC LIMIT 10")
     List<UnknownWord> findTop10ByUserIdAndNextReviewDateBeforeOrEqual(@Param("userId") Long userId, @Param("today") LocalDate today);
